Print subfolder names even in non-recursive mode
When hivetool recursively prints from hiveconf it first prints the
parameters of the start-level and then the sub-folder name followed by
its contents for each lower level.

Note that the foldername-variable in first iteration of the loop in
print_walk() is always equal to the root-parameter. This means that
the first foldername code will only run in recursive mode.

When printing non-recursively we didn't print the sub-folder names. This
was a regression caused by the Python 3 conversion.

Signed-off-by: Samuel Mannehed <samuel@cendio.se>