Handle trailing slash folders in recursive mode
In recursive mode, the folderpath would end with a "/" which would, in
the old code result in double slashes in the subfolder path. This in
turn led to the correct folder not being found, giving us a crash. This
was once again fixed by using os.path.join(). Should have been part of
commit 20e1b47bf87c787f6d8bd9a48610345c78f98144.

Signed-off-by: Samuel Mannehed <samuel@cendio.se>